# specs/001-pilot-space-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pilot Space API
  description: |
    AI-augmented SDLC platform API for project management with intelligent features.

    ## Authentication
    All endpoints require Bearer token authentication via Supabase Auth (GoTrue).
    Authorization is enforced via Row-Level Security (RLS) policies.

    ## Rate Limits
    - Standard endpoints: 1000 requests/minute per workspace
    - AI endpoints: 100 requests/minute per workspace (subject to LLM provider limits)
    - Webhook endpoints: 500 requests/minute

    ## Pagination
    List endpoints support cursor-based pagination using `cursor` and `limit` parameters.
    Response format: `{ data: [...], meta: { total, cursor, hasMore } }`

    ## Updated
    Session 2026-01-22: Migrated to Supabase platform
  version: 2.0.0
  contact:
    name: Pilot Space Team
    url: https://pilot-space.com
  license:
    name: MIT
servers:
  - url: https://api.pilot-space.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
...
paths:
...
  /workspaces/{workspaceSlug}/projects/{projectId}/issues:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [Issues]
      summary: List issues in project
      operationId: listIssues
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/Priority"
        - name: assigneeId
          in: query
          schema:
            type: string
            format: uuid
        - name: cycleId
          in: query
          schema:
            type: string
            format: uuid
        - name: moduleId
          in: query
          schema:
            type: string
            format: uuid
        - name: labels
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIssues"

    post:
      tags: [Issues]
      summary: Create issue
      operationId: createIssue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /workspaces/{workspaceSlug}/projects/{projectId}/issues/{issueId}:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/issueId"
    get:
      tags: [Issues]
      summary: Get issue detail
      operationId: getIssue
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Issue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetail"

    patch:
      tags: [Issues]
      summary: Update issue
      operationId: updateIssue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueRequest"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetail"

    delete:
      tags: [Issues]
      summary: Delete issue (soft delete)
      operationId: deleteIssue
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Issue deleted
  /workspaces/{workspaceSlug}/projects/{projectId}/issues/{issueId}/activities:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/issueId"
    get:
      tags: [Issues]
      summary: Get issue activity log
      operationId: getIssueActivities
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Activity log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedActivities"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceSlug:
      name: workspaceSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]+$"
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    issueId:
      name: issueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    cursor:
      name: cursor
      in: query
      schema:
        type: string
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 100
  schemas:
...
    Priority:
      type: string
      enum: [urgent, high, medium, low, none]
    StateGroup:
      type: string
      enum: [backlog, unstarted, started, completed, cancelled]

    # ==================== User ====================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Issue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequenceId:
          type: integer
        identifier:
          type: string
          description: "Project identifier + sequence (e.g., PS-123)"
        name:
          type: string
        description:
          type: string
        state:
          $ref: "#/components/schemas/State"
        priority:
          $ref: "#/components/schemas/Priority"
        assignee:
          $ref: "#/components/schemas/User"
        reporter:
          $ref: "#/components/schemas/User"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        estimatePoints:
          type: integer
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date
        cycleId:
          type: string
          format: uuid
        moduleId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IssueDetail:
      allOf:
        - $ref: "#/components/schemas/Issue"
        - type: object
          properties:
            descriptionHtml:
              type: string
            subIssues:
              type: array
              items:
                $ref: "#/components/schemas/Issue"
            integrationLinks:
              type: array
              items:
                $ref: "#/components/schemas/IntegrationLink"
            noteLinks:
              type: array
              items:
                $ref: "#/components/schemas/NoteLink"
            aiMetadata:
              type: object
...
    NoteLink:
      type: object
      properties:
        noteId:
          type: string
          format: uuid
        title:
          type: string
        blockId:
          type: string
    CreateIssueRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        stateId:
          type: string
          format: uuid
        priority:
          $ref: "#/components/schemas/Priority"
        assigneeId:
          type: string
          format: uuid
        labelIds:
          type: array
          items:
            type: string
            format: uuid
        estimatePoints:
          type: integer
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date
        cycleId:
          type: string
          format: uuid
        moduleId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
    UpdateIssueRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        stateId:
          type: string
          format: uuid
        priority:
          $ref: "#/components/schemas/Priority"
        assigneeId:
          type: string
          format: uuid
        labelIds:
          type: array
          items:
            type: string
            format: uuid
        estimatePoints:
          type: integer
        startDate:
          type: string
          format: date
        targetDate:
          type: string
          format: date
        cycleId:
          type: string
          format: uuid
        moduleId:
          type: string
          format: uuid
    PaginatedIssues:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        cursor:
          type: string
        hasMore:
          type: boolean
    State:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        group:
          $ref: "#/components/schemas/StateGroup"
    Label:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
    IntegrationLink:
      type: object
      properties:
        id:
          type: string
          format: uuid
        integrationType:
          type: string
          enum: [github]
        linkType:
          type: string
          enum: [pr, commit, branch]
        externalId:
          type: string
        externalUrl:
          type: string
          format: uri
        metadata:
          type: object
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        verb:
          type: string
        field:
          type: string
        oldValue:
          type: string
        newValue:
          type: string
        comment:
          type: string
        commentHtml:
          type: string
        actor:
          $ref: "#/components/schemas/User"
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PaginatedActivities:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        cursor:
          type: string
        hasMore:
          type: boolean

    # ==================== AI ====================
...